+++ src/modules/dashboard/helpers/parse-object-rutas.ts
@@ -32,7 +32,6 @@ export const parseObjectRutas = (rutas: RutaInterface[], router: Router) => {
             rutasHijas: rutasHijasParseadas
         }
     })
-    console.log(rutasParseadas)
     return rutasParseadas
 }
