+++ src/modules/expediente/components/CustomAccordion.vue
@@ -1,15 +1,15 @@
 <template>
   <div>
     <div class="border-b border-gray-200 dark:border-gray-700">
-      <h2 id="accordion-heading-1">
+      <h2 :id="headingId">
         <button 
           type="button" 
           @click="toggle" 
           :aria-expanded="isOpen ? 'true' : 'false'"
-          aria-controls="accordion-body-1"
+          :aria-controls="bodyId"
           class="flex items-center justify-between w-full p-3 font-medium text-gray-500 border-b border-gray-200 rounded-t-lg focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-800 dark:border-gray-700 dark:text-gray-400 hover:bg-gray-100 dark:hover:bg-gray-800 gap-2"
         >
-          <span>Detalles del Expediente</span>
+          <span>{{ title }}</span>
           <svg 
             class="w-3 h-3 transition-transform duration-300" 
             :class="{'rotate-180': isOpen}" 
@@ -29,8 +29,8 @@
         </button>
       </h2>
       <div 
-        id="accordion-body-1" 
-        aria-labelledby="accordion-heading-1" 
+        :id="bodyId" 
+        :aria-labelledby="headingId" 
         v-show="isOpen" 
         class="p-3 border border-b-0 border-gray-200 dark:border-gray-700"
       >
@@ -41,9 +41,17 @@
 </template>
 
 <script setup lang="ts">
-import { ref } from 'vue';
+import { ref, defineProps } from 'vue';
+
+const props = defineProps<{
+  title?: string;
+}>();
+
+const title = props.title; // Usar props directamente
 
 const isOpen = ref(false);
+const headingId = 'accordion-heading-' + Math.random().toString(36).substr(2, 9);
+const bodyId = 'accordion-body-' + Math.random().toString(36).substr(2, 9);
 
 const toggle = () => {
   isOpen.value = !isOpen.value;
