+++ src/modules/expediente/pages/AgregarPage.vue
@@ -1,13 +1,14 @@
 <template>
-  <div>
-    <p>Agregar</p>
-  </div>
+  <DashboardLayout>
+      <h1>Te damos la bienvenida al módulo de Expediente</h1>
+      <h2>Un módulo del sistema de gestión documental</h2>
+  </DashboardLayout>
 </template>
 
 <script setup lang="ts">
+import DashboardLayout from '@/modules/dashboard/layouts/DashboardLayout.vue';
+
 
 </script>
 
-<style scoped>
-
-</style>
+<style scoped></style>
