--- src/modules/expediente/pages/IntroductionPage.vue
@@ -1,14 +1,0 @@
-<template>
-  <DashboardLayout>
-      <h1>Te damos la bienvenida al módulo de Expediente</h1>
-      <h2>Un módulo del sistema de gestión documental</h2>
-  </DashboardLayout>
-</template>
-
-<script setup lang="ts">
-import DashboardLayout from '@/modules/dashboard/layouts/DashboardLayout.vue';
-
-
-</script>
-
-<style scoped></style>
